:recycle: refactor api client
